Move all common tests into cpp files from hpp
