--- common/test/src/util/string_test.hpp
@@ -1,20 +1,0 @@
-//
-// Created by X-ray on 05/04/2025.
-//
-
-#ifndef STRING_TEST_HPP_04234325
-#define STRING_TEST_HPP_04234325
-#include <base-common/util/string.hpp>
-#include <gtest/gtest.h>
-
-TEST(String, ReplaceAll) {
-    const std::string str = "Hello, world! Hello, world!";
-    const std::string target_replace = "Hello";
-    const std::string replace_val = "Hi";
-
-    const std::string result = base::common::util::string::ReplaceAll(str, target_replace, replace_val);
-
-    ASSERT_EQ(result, "Hi, world! Hi, world!");
-}
-
-#endif //STRING_TEST_HPP_04234325
